1.08   pair down bundle details.        fix defaults.plist; fix drawers their defaults.        ignoring device power up strings - and use        it to ask/trigger a #C0 5 seconds later.        tooltips added.        make graph backgrounds properly white.        WiFi now shows channel # rather than Ghz        sd-ish spread around values        dBm values corrected for narrow scales        bettert serial handling        all strings localized        fix Hz on max - sensible range.        more tooltips        fixed paused scaling live display.

+++ iRfExplorer/iRFExplorer/NumericScaleDefinition.h
@@ -42,7 +42,7 @@
 	     withDiff:(double)_diff
  withNiceNbrScore:(double)_niceNbrScore
    withNbrOfTicks:(int)_nbrOfTicks
- withSubTickCount:(double)_subCount
+ withSubTickCount:(int)_subCount
 	 includesZero:(BOOL)_includesZero
  withDataCoverage:(double)_dataCoverage;
